+++ main/src/cgeo/geocaching/sorting/FindsComparator.java
@@ -31,7 +31,7 @@ public class FindsComparator extends AbstractCacheComparator {
         }
         Integer logged = cache.getLogCounts().get(LogType.LOG_FOUND_IT);
         if (logged != null) {
-            finds = logged.intValue();
+            finds = logged;
         }
         return finds;
     }
